Moving all function directories under cmd directory
