+++ cmd/httpTrigger/trigger.go
@@ -1,7 +1,7 @@
 package httpTrigger
 
 import (
-	"github.com/ind9/rocks/ops"
+	"github.com/ind9/rocks/cmd/ops"
 	"github.com/spf13/cobra"
 )
