+++ cmd/ops/rocks.go
