+++ cmd/restore/restore.go
@@ -8,7 +8,7 @@ import (
 	"runtime"
 
 	"github.com/hashicorp/go-multierror"
-	"github.com/ind9/rocks/ops"
+	"github.com/ind9/rocks/cmd/ops"
 	"github.com/spf13/cobra"
 	"github.com/tecbot/gorocksdb"
 )
