+++ cmd/statistics/statistics.go
@@ -9,7 +9,7 @@ import (
 	"sync"
 
 	"github.com/hashicorp/go-multierror"
-	"github.com/ind9/rocks/ops"
+	"github.com/ind9/rocks/cmd/ops"
 	"github.com/spf13/cobra"
 	"github.com/tecbot/gorocksdb"
 )
